Add hardware detection to choose between Torch and MLX
