+++ rag/file_conversion_router/conversion/pdf_converter.py
@@ -1,22 +1,19 @@
+import os
+import re
 import subprocess
 from pathlib import Path
-import os
+
 import fitz
-import re
 from pix2text import Pix2Text
 
 from rag.file_conversion_router.conversion.base_converter import BaseConverter
-from rag.file_conversion_router.utils.hardware_detection import detect_gpu_setup
+from rag.file_conversion_router.services.tai_nougat_service import TAINougatConfig
 from rag.file_conversion_router.services.tai_nougat_service.api import convert_pdf_to_mmd
-from rag.file_conversion_router.services.tai_nougat_service.nougat_config import NougatConfig
 
 
 class PdfConverter(BaseConverter):
     def __init__(self):
         super().__init__()
-        self.device_type, _ = detect_gpu_setup()  # Detect hardware configuration
-
-        self._logger.info(f"Using {self.device_type} on Torch")
 
     def convert_pdf_to_markdown(self, pdf_file_path, output_file_path, page_numbers=None):
         """
@@ -129,6 +126,7 @@ class PdfConverter(BaseConverter):
         # Convert the PDF to Markdown using Nougat.
         # self._to_markdown_using_native_nougat_cli(pdf_without_images_path, output_path)
         self._to_markdown_using_tai_nougat(pdf_without_images_path, output_path)
+        # self._to_markdown_using_mlx_nougat(pdf_without_images_path, output_path)
 
         # Now change the file name of generated mmd file to align with the expected md file path from base converter
         output_mmd_path = output_path.with_suffix(".mmd")
@@ -138,7 +136,6 @@ class PdfConverter(BaseConverter):
         output_mmd_path.rename(target)
         print(output_mmd_path)
         return target
-
 
     def _to_markdown_using_native_nougat_cli(self, input_pdf_path: Path, output_path: Path) -> None:
         """
@@ -167,15 +164,13 @@ class PdfConverter(BaseConverter):
             self._logger.error(f"An error occurred: {str(e)}")
             raise
 
+    @staticmethod
+    def _to_markdown_using_tai_nougat(input_pdf_path: Path, output_path: Path) -> None:
+        """Perform PDF to Markdown conversion using TAI Nougat.
 
-    def _to_markdown_using_tai_nougat(self, input_pdf_path: Path, output_path: Path) -> None:
+        TAI nougat is our custom implementation of the Nougat API, with better performance and abstraction.
         """
-        Perform PDF to Markdown conversion using TAI Nougat.
-
-        TAI nougat is our custom implementation of the Nougat API, with better abstraction
-        and especially optimization on avoiding loading nougat model repetitively.
-        """
-        config = NougatConfig(
+        config = TAINougatConfig(
             pdf_paths=[input_pdf_path],
             output_dir=output_path.parent,
         )
