+++ rag/file_conversion_router/services/tai_nougat_service/torch_nougat_service/README.md
@@ -1,0 +1,5 @@
+# Torch Nougat Service
+
+We have migrated code from the Meta Nougat repository to enable more flexible development of Nougat services, including support for **Dependency Injection**.
+
+The original Meta Nougat is built on Torch.
